+++ src/projectgoldstarscalculators/ProjectGoldStarsCalculatorS.java
@@ -4,8 +4,8 @@ import java.awt.Font;
 import java.io.File;
 public class ProjectGoldStarsCalculatorS
 {
-    public static final String NAME = "Project GoldStars Calculator S 2.3";
-    public static final String VERSION = "Version 2.3";
+    public static final String NAME = "Project GoldStars Calculator S 2.3 Update 1";
+    public static final String VERSION = "Version 2.3.1";
     public static final File FOLDER = new File(System.getProperty("user.home"), "Project GoldStars Calculator S");
     public static Color color1;
     public static Color color2;