+++ Changelog.md
@@ -2,6 +2,11 @@
 ### Changelog
 
 ## Version 2.3.x
+### Version 2.3.1
+* Updated the instructions in the Miles Per Gallon Calculator.
+* Updated the instructions in the Pythagorean Triple Finder.
+* Other minor changes.
+
 ### Version 2.3
 * Added instructions to the Logarithms window.
 * Added instructions to the Square & Cube Roots window.